binance_tracker/requirements.txt
python-binance
customtkinter
cryptography